requirements.txt
flask==3.0.3
python-dotenv==1.0.1
numpy==1.26.4
Pillow==10.4.0
matplotlib==3.9.2
torch==2.4.0
pytest==8.3.2
